+++ bin/build_config
@@ -12,11 +12,11 @@ require 'rubocop/rspec/config_formatter'
 glob = File.join('lib', 'rubocop', 'cop', 'rspec',
                  '{,capybara,factory_bot,rails}', '*.rb')
 # Due to YARD's sensitivity to file require order (as of 0.9.25),
-# we have to prepend the list with our base cop, RuboCop::Cop::RSpec::Cop.
+# we have to prepend the list with our base cop, RuboCop::Cop::RSpec::Base.
 # Otherwise, cop's parent class for cops loaded before our base cop class
-# are detected as RuboCop::Cop::Cop, and that complicates the detection
+# are detected as RuboCop::Cop::Base, and that complicates the detection
 # of their relation with RuboCop RSpec.
-rspec_cop_path = File.join('lib', 'rubocop', 'cop', 'rspec', 'cop.rb')
+rspec_cop_path = File.join('lib', 'rubocop', 'cop', 'rspec', 'base.rb')
 YARD.parse(Dir[glob].prepend(rspec_cop_path), [])
 
 descriptions = RuboCop::RSpec::DescriptionExtractor.new(YARD::Registry.all).to_h
