+++ lib/rubocop-rspec.rb
@@ -19,6 +19,7 @@ require_relative 'rubocop/rspec/example_group'
 require_relative 'rubocop/rspec/example'
 require_relative 'rubocop/rspec/hook'
 require_relative 'rubocop/rspec/variable'
+require_relative 'rubocop/cop/rspec/base'
 require_relative 'rubocop/cop/rspec/cop'
 require_relative 'rubocop/rspec/align_let_brace'
 require_relative 'rubocop/rspec/factory_bot'
