+++ spec/project/default_config_spec.rb
@@ -25,7 +25,7 @@ RSpec.describe 'config/default.yml' do
         "#{namespace}/#{cop_name}"
       end
 
-    cop_names - %w[RSpec/Cop]
+    cop_names - %w[RSpec/Cop RSpec/Base]
   end
 
   let(:config_keys) do
